requirements.txt
fastapi
uvicorn
jinja2
requests
beautifulsoup4
sqlalchemy
psycopg2-binary
python-dotenv
python-dateutil
feedparser
aiohttp
asyncio
